Add explicit cast when computing the available amount of system memory.

+++ tests/src/runtimeApi/memory/hipMemcpy.cpp
@@ -292,8 +292,8 @@ void memcpytest2_get_host_memory(size_t& free, size_t& total) {
     MEMORYSTATUSEX status;
     status.dwLength = sizeof(status);
     GlobalMemoryStatusEx(&status);
-    free = (0.4 * status.ullAvailPhys);
-    total = (0.4 * status.ullTotalPhys);
+    free = static_cast<size_t>(0.4 * status.ullAvailPhys);
+    total = static_cast<size_t>(0.4 * status.ullTotalPhys);
 }
 #else
 struct sysinfo memInfo;
